Auto config via domain from .env and add self documentation

+++ .docker/bin/generate_certs.sh
@@ -1,7 +1,15 @@
 #! /bin/bash
 # Doc https://github.com/FiloSottile/mkcert
-sudo apt install libnss3-tools
-wget -O mkcert https://github.com/FiloSottile/mkcert/releases/download/v1.4.4/mkcert-v1.4.4-linux-amd64
-sudo chmod +x ./mkcert
+
+FILE=./mkcert
+if [ ! -f "$FILE" ]; then
+    sudo apt install libnss3-tools
+    wget -O mkcert "https://github.com/FiloSottile/mkcert/releases/download/v$2/mkcert-v$2-linux-amd64"
+    sudo chmod +x ${FILE}
+    sudo chown -R $USER:$USER ${FILE}
+else
+    ./mkcert -uninstall
+fi
+
 ./mkcert -install
-./mkcert -cert-file ./../config/ssl/server.cert -key-file ./../config/ssl/server.key sefu.test "*.sefu.test"
+./mkcert -cert-file ./../config/ssl/server.cert -key-file ./../config/ssl/server.key $1 "*.$1"
